+++ test/tutorials/src/typescript.ts
@@ -99,6 +99,17 @@ async function run() {
         statics: heroCollectionMethods
     });
 
+    // add a preInsert-hook
+    myDatabase.heroes.postInsert(
+        function myPostInsertHook(
+            this: HeroCollection, // own collection is bound to the scope
+            docData, // documents data
+            doc // RxDocument
+        ) {
+            console.log('insert to ' + this.name + '-collection: ' + doc.firstName);
+        },
+        false // not async
+    );
 
 
     /**